% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{add_transform}
\alias{add_transform}
\title{Update a formula with provided transform functions.}
\usage{
add_transform(x, transforms = NULL, lag = NULL)
}
\arguments{
\item{x}{The variables to add transformations to.}

\item{transforms}{A dataframe identifying variables and transforms to use for each variable.}

\item{lag}{The lag level to remove from the formula.}
}
\value{
An updated formula with appropriate variable transformation functions added.
}
\description{
This internal function takes a formula and modifies it to include the
variable transforms provided as an input. This function is used in the 'create_new_formula'
function.
}
\seealso{
Other research plan functions: 
\code{\link{analyze_plan}()},
\code{\link{build_other_values}()},
\code{\link{create_new_formula}()},
\code{\link{create_plan_data}()},
\code{\link{create_research_formulas}()},
\code{\link{create_variable_table}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_problematic_variables}()},
\code{\link{identify_variable_transformations}()},
\code{\link{research_plan}()},
\code{\link{trim_formula}()}
}
\concept{research plan functions}
