% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{analysis}
\alias{analysis}
\title{Main entry function to conduct analysis of a dataset.}
\usage{
analysis(
  runs,
  formula,
  data,
  cluster = NULL,
  weights = NULL,
  model.type = NULL,
  model.extra.args = NULL,
  inference = c("frequentist", "bayesian")
)
}
\arguments{
\item{runs}{The number of runs to conduct. Optionally, an object returned from "resample_df."}

\item{formula}{The formula to run.}

\item{data}{The data to use for analysis.}

\item{cluster}{Formula specifying the variable in data that identifies clusters within the data.}

\item{weights}{Vector of weights to use.}

\item{model.type}{The model type is determined automatically from the formula and data. Can optionally be specified directly.}

\item{model.extra.args}{Extra arguments to the model that will be run.}

\item{inference}{The type of inference to use: frequentist (the default) or bayesian.}
}
\description{
This function runs either a frequentist or bayesian analysis of a dataset.
}
\examples{
analysis(runs = 1000, formula = out ~ treat, data = dt, inference = "bayesian")

}
\keyword{analysis}
\keyword{bayesian}
\keyword{boostrap}
\keyword{frequentist}
