% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{analyze_plan}
\alias{analyze_plan}
\title{Function to automatically analyze a research plan in a systematic and efficient manner.}
\usage{
analyze_plan(
  research.plan,
  .threshold = 0.3,
  only.run = NULL,
  fill.in.plan = NULL,
  run.basic = T,
  run.mediation = T,
  run.interaction = T
)
}
\arguments{
\item{research.plan}{A research plan to analyze. This should be the return value from the function 'research_plan'.}

\item{.threshold}{The p-value used to identify which variables correlate with the treatment and thus which potential
confounding variables are included in the analysis.}

\item{only.run}{The name of a variable in the research plan to analyze. If this is provided only the named variable will
be analyzed and returned,}

\item{fill.in.plan}{An optional results from a previous 'analyze_plan' run. If provided, the function will ignore values
already present and just run variables, interactions, or mediators that are not present. Using this option is an easy way
to add a new variable to a plan and run just this new variable.}

\item{run.basic}{Logical indicator of whether to run the main effects for a variable.}

\item{run.mediation}{Logical indicator of whether to run the interaction effects for a variable.}

\item{run.interaction}{Logical indicator of whether to run the mediation effects for a variable.}
}
\value{
The statistical results of the analysis. Each list item corresponds to a variable in the research plan. Within each
  variable separate dataframes contain the results for the main analysis, the interaction effects, and the mediation analysis.
  If there are multiple outcomes examined these are added to the dataframe. If time is a component (e.g., when conducting
  survival anlysis) time is also added to the dataframe.

  Mediation analysis is only performed if the main result for a variable/outcome is significant.
}
\description{
Takes a research plan created from the function 'research_plan' and performs all necessary analysis. This analysis is then
returned to the user to allow various types of assessment.
}
\seealso{
Other research plan functions: 
\code{\link{add_transform}()},
\code{\link{build_other_values}()},
\code{\link{create_new_formula}()},
\code{\link{create_plan_data}()},
\code{\link{create_research_formulas}()},
\code{\link{create_variable_table}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_problematic_variables}()},
\code{\link{identify_variable_transformations}()},
\code{\link{research_plan}()},
\code{\link{trim_formula}()}
}
\concept{research plan functions}
\keyword{analysis,}
\keyword{automated,}
\keyword{plan}
\keyword{research}
