% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{build_other_values}
\alias{build_other_values}
\title{Automates creation of variables in formula.}
\usage{
build_other_values(
  .data,
  ...,
  unit,
  prefix = "other",
  func = base::sum,
  .func.args = NULL,
  .other = T
)
}
\arguments{
\item{.data}{The dataframe containing all variables.}

\item{...}{Variables for which to create "other" values. Should be entered a symbols and not strings.}

\item{unit}{A vector listing the variables that describe the unit of analysis in the data.}

\item{prefix}{A string indicating the string to be added as a prefix to the variable name to indicate that
it is an "other" variable.}

\item{func}{The function applied to the vector of valuables in each unit of analysis.}

\item{.func.args}{Arguments to pass to 'func'.}

\item{.other}{Whether or not to subtract observations in each row from the "other" observations in each unit.
The default is to do so, i.e., if you want to sum the number of other attacks against a target the number of
attacks initiated by the source in the current observation is subtracted.}
}
\value{
A dataframe with the other variables created. The column names are the names of the variables. The number of
  of rows should match the numbr of rows in'.data'.
}
\description{
This function is a helper function to automatically create variables (in both a formula and the data)
that identify other "actions" taken against a target in a unit of analysis. Specifically, it identifies
the "func" (defaults to "sum") of the variables for each observation of a unit minus the current values
for that unit.
}
\details{
The goal is to make this process as generalizable as possible so these functions are usable in a variety
of contexts.
}
\seealso{
Other research plan functions: 
\code{\link{add_transform}()},
\code{\link{analyze_plan}()},
\code{\link{create_new_formula}()},
\code{\link{create_plan_data}()},
\code{\link{create_research_formulas}()},
\code{\link{create_variable_table}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_problematic_variables}()},
\code{\link{identify_variable_transformations}()},
\code{\link{research_plan}()},
\code{\link{trim_formula}()}
}
\concept{research plan functions}
\keyword{formulas,}
\keyword{variables}
