% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{collapse_time}
\alias{collapse_time}
\title{Summarize a dataframe to a different unit of analysis.}
\usage{
collapse_time(data, time.var, time.factor, group.by, variables)
}
\description{
Summarize a dataframe to a different unit of analysis.
}
