% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{create_model_string}
\alias{create_model_string}
\title{Create a string describing a model run.}
\usage{
create_model_string(
  model.type,
  formula.parsed,
  cluster,
  weights,
  inference,
  has.tve
)
}
\arguments{
\item{model.type}{The type of model being run.}

\item{formula.parsed}{The parsed formula object.}

\item{cluster}{The clusters used to produce the bootstrap test data.}

\item{weights}{The weights used to produce the bootstrap test data.}

\item{inference}{Whether the model will be run using frequentist or Bayesian inference.}

\item{has.tve}{Whether the event history model has time-varying hazards.}
}
\value{
A formatted string that describes the model being run.
}
\description{
This internal function takes information about the model being run and turns it into a
well formatted string.
}
