% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{create_new_formula}
\alias{create_new_formula}
\title{Create a new 'planformula' object.}
\usage{
create_new_formula(
  dv = NULL,
  iv,
  controls,
  type = "*",
  transforms = NULL,
  lag = NULL
)
}
\arguments{
\item{dv}{A character string naming the dependent variable.}

\item{iv}{A vector of character strings naming the main independent variables.}

\item{controls}{A vector of character strings naming the control variables.}

\item{type}{A character indicating the relationship between the 'iv's.}

\item{transforms}{A list providing variable names and transformation for those variables.}

\item{lag}{Lags to include (currently unused).}
}
\value{
An object that includes all the components (including variable transformations) for a
testable formula. This formula object is passed to other functions to analyze.
}
\description{
This internal function takes components of a formula and creates a new formula object
that is used by other research plan functions. The 'planformula' object also has its
own print and update functions.
}
\seealso{
Other research plan functions: 
\code{\link{add_transform}()},
\code{\link{analyze_plan}()},
\code{\link{build_other_values}()},
\code{\link{create_plan_data}()},
\code{\link{create_research_formulas}()},
\code{\link{create_variable_table}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_problematic_variables}()},
\code{\link{identify_variable_transformations}()},
\code{\link{research_plan}()},
\code{\link{trim_formula}()}
}
\concept{research plan functions}
