% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{create_plan_data}
\alias{create_plan_data}
\title{Create all data that corresponds to newly created variables.}
\usage{
create_plan_data(data, all.vars, lag, prefix, unit, keep.vars)
}
\arguments{
\item{data}{Data to use.}

\item{all.vars}{A list of new variables to create. This is the first list item returned from 'create_variable_table'.}

\item{lag}{The number of lags to include.}

\item{prefix}{A list that provides the prefixes for source, target, relative, and other.}

\item{unit}{A vector of strings identifying the name of the unit of analysis.}

\item{keep.vars}{Additional variables to keep in 'data' when returned.}
}
\value{
A dataframe that includes all of the new variables created. This dataframe will also include the variables
  identified by 'keep.vars'.
}
\description{
Given variables specified in 'create_variable_table' this function creates the new corresponding variables
in the data.
}
\seealso{
Other research plan functions: 
\code{\link{add_transform}()},
\code{\link{analyze_plan}()},
\code{\link{build_other_values}()},
\code{\link{create_new_formula}()},
\code{\link{create_research_formulas}()},
\code{\link{create_variable_table}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_problematic_variables}()},
\code{\link{identify_variable_transformations}()},
\code{\link{research_plan}()},
\code{\link{trim_formula}()}
}
\concept{research plan functions}
\keyword{formula,}
\keyword{selection}
\keyword{variable,}
