% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{create_research_formulas}
\alias{create_research_formulas}
\title{Create formulas for testing.}
\usage{
create_research_formulas(all.vars, all.transform, lag)
}
\arguments{
\item{all.vars}{The return value from the function 'create_variable_table'.}

\item{all.transform}{The return value from the function 'identify_variable_transformations'.}

\item{lag}{The number of lags for each control variable to include.}
}
\value{
A list of formulas and predictions that allow main effects, interaction effects, and mediating effects to be
  tested.
}
\description{
Given suggested variables to include and how to transform them, this function creates the actual formulas that
can be tested in a regression model.
}
\details{
For each variable identified as a treatment this function creates a formula for the main effect, a formula for each
interaction, and formulas for each possible mediator. For each formula, a prediction to test the formula is also created.
}
\seealso{
Other research plan functions: 
\code{\link{add_transform}()},
\code{\link{analyze_plan}()},
\code{\link{build_other_values}()},
\code{\link{create_new_formula}()},
\code{\link{create_plan_data}()},
\code{\link{create_variable_table}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_problematic_variables}()},
\code{\link{identify_variable_transformations}()},
\code{\link{research_plan}()},
\code{\link{trim_formula}()}
}
\concept{research plan functions}
\keyword{formula,}
\keyword{transformation}
\keyword{variable,}
