% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{create_stan_fit}
\alias{create_stan_fit}
\title{Create 'stanfit' object.}
\usage{
create_stan_fit()
}
\arguments{
\item{model}{The model.}
}
\value{
Return a 'stanfit' s4 object.
}
\description{
This internal function is in progress.
}
