% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{create_stan_summary}
\alias{create_stan_summary}
\title{Create a stan summary for summary/print output.}
\usage{
create_stan_summary(full.matrix, full.names)
}
\arguments{
\item{full.matrix}{A matrix with all variables and values for each bootstrap run.}

\item{full.names}{The list of names that correspond to columns in 'full.matrix'.}
}
\value{
A formatted summary matrix for display.
}
\description{
This internal function takes the full bootstrap return matrix/names and outputs a formatted stan
summary.
}
