% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{create_values}
\alias{create_values}
\title{Create high and low values}
\usage{
create_values(x, .quantile = c(0.975, 0.025), .places = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{.quantile}{The high and low quantiles for a numeric vector.}

\item{.places}{The number of places to round a numeric vector to.}
}
\value{
For a numeric vector this function return high and low values.
  For a character vector this returns unique values.
}
\description{
This function takes in a vector (numeric or character) and outputs nicely formatted
test values.
}
