% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{create_variable_table}
\alias{create_variable_table}
\title{Create all the new variables to include in each formula.}
\usage{
create_variable_table(
  data,
  treatment,
  control,
  interaction,
  mediation,
  factor.transform,
  lag,
  unit,
  prefix
)
}
\arguments{
\item{data}{Data to use.}

\item{treatment}{The list of treatment variables to check. May have two names in the list: action and factor.
An 'action' is a variable that can have source/target/other transforms. A 'factor' is not transformed.}

\item{control}{The list of control variables.}

\item{interaction}{The list of interaction variables.}

\item{mediation}{The list of mediation variables.}

\item{factor.transform}{A provided function that allows a factor/character variable to be transformed.}

\item{lag}{The number of lags to include.}

\item{unit}{A list identifying the unit of analysis in the data. May be NULL if no unit is present.}

\item{prefix}{A list that provides the prefixes for source, target, relative, and other.}
}
\value{
A list with two values. The first value is a dataframe that lists all of the variables created and their
  transformations. The second provides labels for all included variables.
}
\description{
This function creates a large number of new variables to potentially include in a formula. These variables
include source values, target values, relative values, and other values.
}
\details{
For included variable lists (treatment, control, interaction, and mediation) the name provides a human-understandable
label while the value gives the name as found in 'data'.
}
\seealso{
Other research plan functions: 
\code{\link{add_transform}()},
\code{\link{analyze_plan}()},
\code{\link{build_other_values}()},
\code{\link{create_new_formula}()},
\code{\link{create_plan_data}()},
\code{\link{create_research_formulas}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_problematic_variables}()},
\code{\link{identify_variable_transformations}()},
\code{\link{research_plan}()},
\code{\link{trim_formula}()}
}
\concept{research plan functions}
\keyword{formula,}
\keyword{selection}
\keyword{variable,}
