% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{ct_to_out}
\alias{ct_to_out}
\title{Nice model output.}
\usage{
ct_to_out(m, cluster = NULL, drop.factor = T, .level = 0.95, .round = 3)
}
\arguments{
\item{m}{The model to print a summary for.}

\item{cluster}{An optional formula identifying variables to use for clustered standard errors.}

\item{drop.factor}{Drop factor variables from print output or keep them. Defaults to drop.}

\item{.level}{Level of statistical significance for confidence intervals. Defaults to P < 0.05.}

\item{.round}{Number of digits to round results to. Defaults to three.}
}
\value{
For a numeric vector this function return high and low values.
  For a character vector this returns unique values.
}
\description{
This function prints out a nice models summary and optionally incorporates clustered standard errors.
}
