% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_info.r
\name{determine_model}
\alias{determine_model}
\title{Determine the type of model to run.}
\usage{
determine_model(formula.parsed, data)
}
\arguments{
\item{formula.parsed}{A parsed formula object returned from the function 'parse_formula'.}

\item{data}{The raw data that the formula will be run on.}
}
\value{
A character vector that identifies the type of model to be run based on the formula and data.
}
\description{
This internal function takes the parsed formula and raw data and determines the type of
model that should be run.
}
\seealso{
Other model setup functions: 
\code{\link{identify_tve}()},
\code{\link{parse_formula}()},
\code{\link{produce_model_function}()}
}
\concept{model setup functions}
