% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{fast_bh}
\alias{fast_bh}
\title{Fast identification of baseline hazard.}
\usage{
fast_bh(model)
}
\arguments{
\item{model}{A survival model running using the "survival" package.}
}
\value{
A list with the baseline hazard and time.
}
\description{
This function takes in a survival model and outputs a list with the
baseline hazard and time.
}
