% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{formula.planformula}
\alias{formula.planformula}
\title{Formula method for planformula}
\usage{
\method{formula}{planformula}(x, to.drop = NULL, transform.iv = NULL)
}
\description{
Formula method for planformula
}
