% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{get_coef}
\alias{get_coef}
\title{Get model coefficients}
\usage{
get_coef(object)
}
\value{
A vector of model coefficients
}
\description{
Get model coefficients
}
\seealso{
Other internal functions: 
\code{\link{get_data}()},
\code{\link{get_draws}()},
\code{\link{get_formula}()},
\code{\link{get_link}()},
\code{\link{get_vcov}()},
\code{\link{results_coefficients}()},
\code{\link{results_contrasts}()},
\code{\link{results_predictions}()},
\code{\link{structure_predictions}()}
}
\concept{internal functions}
