% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{get_data}
\alias{get_data}
\title{Get data from a frequentist or Bayesian object.}
\usage{
get_data(object, data = NULL)
}
\arguments{
\item{object}{The model object.}
}
\value{
The data used in the model.
}
\description{
This function selects the data from a model object produced by 'analyze'. This function also
works on model objects produced by 'rstanarm', e.g., from a run of 'stan_glm'.
}
\seealso{
Other internal functions: 
\code{\link{get_draws}()},
\code{\link{get_formula}()},
\code{\link{results_coefficients}()},
\code{\link{results_contrasts}()},
\code{\link{results_predictions}()},
\code{\link{structure_predictions}()}
}
\concept{internal functions}
