% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{get_draws}
\alias{get_draws}
\title{Get draws from a frequentist or Bayesian object.}
\usage{
get_draws(object)
}
\arguments{
\item{object}{The model object.}
}
\value{
The dataframe containing all of the draws (model runs).
}
\description{
This function selects the draws dataframe from a model object produced by 'analyze'. This function also
works on model objects produced by 'rstanarm', e.g., from a run of 'stan_glm'.
}
\seealso{
Other internal functions: 
\code{\link{get_data}()},
\code{\link{get_formula}()},
\code{\link{get_inv_link}()},
\code{\link{get_vcov}()},
\code{\link{results_coefficients}()},
\code{\link{results_contrasts}()},
\code{\link{results_predictions}()},
\code{\link{structure_predictions}()}
}
\concept{internal functions}
