% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{get_link}
\alias{get_link}
\title{Get the inverse model link from a frequentist or Bayesian object.}
\usage{
get_link(object)
}
\arguments{
\item{object}{The model object.}
}
\value{
The data used in the model.
}
\description{
Returns the function that allows coefficients to be transformed into a number on the scale of the result.
}
\seealso{
Other internal functions: 
\code{\link{get_coef}()},
\code{\link{get_data}()},
\code{\link{get_draws}()},
\code{\link{get_formula}()},
\code{\link{get_vcov}()},
\code{\link{results_coefficients}()},
\code{\link{results_contrasts}()},
\code{\link{results_predictions}()},
\code{\link{structure_predictions}()}
}
\concept{internal functions}
