% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{get_prediction_frequentist}
\alias{get_prediction_frequentist}
\title{Function to get results from an object returned by a frequentist analysis.}
\usage{
get_prediction_frequentist(
  model,
  predictions = NULL,
  cluster = NULL,
  type = c("observed values", "mean", "zero"),
  response = T,
  fe.fix = T
)
}
\arguments{
\item{model}{A statistical model to get predictions from.}

\item{prediction}{Predictions structured in the normal way.}
}
\description{
A test function to return predictions and contrasts from a frequentist model. Allows
clustered standard errors.
}
\seealso{
Other results: 
\code{\link{qualitative_assessment}()},
\code{\link{results_mediation}()},
\code{\link{results}()},
\code{\link{summarize_interval}()}
}
\concept{results}
\keyword{frequentist}
\keyword{prediction}
\keyword{results}
