% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{get_summary_statistics}
\alias{get_summary_statistics}
\title{Function to get summary statistics from a dataframe.}
\usage{
get_summary_statistics(formula.list, data, labels = NULL)
}
\arguments{
\item{formula.list}{A formula or list of formulas that indicate which variables are included in summary statistics.}

\item{data}{Dataframe with variables to produce summary statistics from.}

\item{labels}{An optional vector of labels: c("variable" = "label).}
}
\value{
A dataframe of summary statistics.
}
\description{
This function produces a table of summary statistics based on a formula and a dataframe.
}
\examples{
get_summary_statistics(formula.list = ~ variable, data = df, labels = c("variable" = "label"))

}
\keyword{summary}
\keyword{summary_stats,}
