% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{get_variable_effects}
\alias{get_variable_effects}
\title{Create predictions for all variables in a formula.}
\usage{
get_variable_effects(
  variables,
  data,
  labels = NULL,
  .variable.set = NULL,
  .perc = c(0.1, 0.9)
)
}
\description{
Create predictions for all variables in a formula.
}
