% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{identify_examples}
\alias{identify_examples}
\title{Identify observations that support our results.}
\usage{
identify_examples(
  results,
  research.plan,
  unit = NULL,
  time = NULL,
  .time.distance = 0.1,
  .quantile = c(0.8, 0.2),
  .variables = c(".main.variable", ".main.interaction")
)
}
\arguments{
\item{results}{The output from 'analyze_plan'.}

\item{research.plan}{The research plan used in the produce of 'results'.}

\item{unit}{The unit of analysis to aggregate examples within.}

\item{time}{The tine variable used in the results. Only necessary for time-series analysis.}

\item{.time.distance}{If a results was from a survival model, this parameter identifies the quantile
of time that indicates a "close" event. If time to the outcome is below this quantile of the time
variable, the observation is considered close.}

\item{.quantile}{The quantile cutoffs for low and high values of a variable. A value above the first
cutoff is considered a "high" value while a value below the second cutoff is considered a "low' value.}

\item{.variables}{Names of the variable columns in the results. Defaults to main and interaction effects.
If results contains more than two variables, additional values need to be added here.}
}
\value{
Function returns a dataframe that contains a supportive observation that matches one of the statistically
  significant results.
}
\description{
This functions takes a set of results (main and interaction but not mediation) and identifies observations
in the dataset used that support the results. Supportive observations are only identified for statistically
significant results.
}
\details{
For most normal regression models, an observation is supportive when (1) it occurs in the same unit of analysis
as a positive outcome and (2) the variable has a high value. For interactions, both the main variable and the interaction
need to have the specified values (conditional effects can occur when the interaction variable has a low value--this is
accounted for). For event history models, a supportive variable does not need to occur in the same observation as the outcome.
Observations that are close in time to the outcome are also considered supportive.
}
\seealso{
Other post analysis exploration: 
\code{\link{get_summary_statistics}()},
\code{\link{get_variable_effects}()}
}
\concept{post analysis exploration}
