% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{identify_problematic_variables}
\alias{identify_problematic_variables}
\title{Exclude variables with insufficient variation to allow inclusion.}
\usage{
identify_problematic_variables(.data, unit.main, variables.to.check)
}
\arguments{
\item{.data}{Data to use.}

\item{unit.main}{A list of variables that identifies the main unit within the data. This is only applicable if there are
multiple observations for each unit (e.g., observations over time for a single actor).}

\item{variables.to.check}{The list of variables to check for sufficient variation.}
}
\value{
This function returns a list of variable names that are problematic to include in regression analysis. These are
  variables that have no variation, no variation within-group variaiton, or that are collinear.
}
\description{
This function takes as input a dataframe, a unit of analysis, and a list of variables and identifies which variables have
sufficient variation to allow inclusion in a regression function. Variables that have insufficient variation in the data,
insufficient variation on average in each unit of analysis, or that are collinear are flagged for exclusion.
}
\details{
This processing step is only applied to numeric variables.
}
\seealso{
Other research plan functions: 
\code{\link{add_transform}()},
\code{\link{analyze_plan}()},
\code{\link{build_other_values}()},
\code{\link{create_new_formula}()},
\code{\link{create_plan_data}()},
\code{\link{create_research_formulas}()},
\code{\link{create_variable_table}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_variable_transformations}()},
\code{\link{research_plan}()},
\code{\link{trim_formula}()}
}
\concept{research plan functions}
\keyword{formula,}
\keyword{selection}
\keyword{variable,}
