% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_info.r
\name{identify_tve}
\alias{identify_tve}
\title{Determine whether a survival model needs to be run with time-varying hazards.}
\usage{
identify_tve(formula, data, var = NULL, random.effects = F)
}
\arguments{
\item{formula}{A formula for the survival model to be run.}

\item{data}{The raw data that the formula will be run on.}

\item{var}{An optional vector indicating which variables to focus on (e.g., when running causal analysis).
If 'var' is not null, non-proportional hazards for remaining variables are ignored.}

\item{random.effects}{A logical variable indicating whether the formula has random effets. If so,
a mixed-effects Cox model is run.}
}
\value{
A data frame with the list of variables (or a subset if 'var' is set) and whether the variable
  has fails the proportional hazards assumption.
}
\description{
This internal function takes a (survival) formula and data and determines if
any variables need to be run with time-varying hazards. Optionally focus only
on 'var' variables.
}
\details{
The basic logic is to run the model conventionally and test the proportional hazards
assumption using 'cox.zph' in the 'survival' package. Test works with normal survival
models run using 'coxph' and mixed-effects models run using 'coxme'.

If time-varying hazards are identified than the appropriate variables are modified in
the formula.
}
\seealso{
Other model setup functions: 
\code{\link{determine_model}()},
\code{\link{parse_formula}()},
\code{\link{produce_model_function}()}
}
\concept{model setup functions}
