% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{identify_variable_transformations}
\alias{identify_variable_transformations}
\title{Based on the data, identify how to transform a variable.}
\usage{
identify_variable_transformations(.data, variables.to.transform)
}
\arguments{
\item{.data}{Data to use.}

\item{variables.to.transform}{The list of variables to identify the appropriate transformation for.}
}
\value{
This function returns a dataframe where each row is a variable and each column characterizes the data of that variable
  and the appropriate transformation function to use. Columns include description of the sign, the portion of minimum and maximum values,
  the skewness, and the kurtosis. Based on these characteristics, an appropriate transformation function for the variable is suggested.
}
\description{
This function takes as input a dataframe and a list of variables and identifies what transformations
should be applied to each variable to make it more normal.
}
\details{
This processing step is only applied to numeric variables.
}
\seealso{
Other research plan functions: 
\code{\link{add_transform}()},
\code{\link{analyze_plan}()},
\code{\link{build_other_values}()},
\code{\link{create_new_formula}()},
\code{\link{create_plan_data}()},
\code{\link{create_research_formulas}()},
\code{\link{create_variable_table}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_problematic_variables}()},
\code{\link{research_plan}()},
\code{\link{trim_formula}()}
}
\concept{research plan functions}
\keyword{formula,}
\keyword{transformation}
\keyword{variable,}
