% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{model_run.frequentist}
\alias{model_run.frequentist}
\title{Run a frequentist model}
\usage{
model_run.frequentist(idx, formula.parsed, model.functions)
}
\arguments{
\item{idx}{A numeric vector of observations in the data to use (for bootstrapping).}

\item{formula.parsed}{The parsed formula object.}

\item{model.functions}{The model object to run on the data.}
}
\value{
A return list formatted to allow compatibility with 'rstanarm'. The list has
  the following items: 'alpha' is the intercept, 'beta' is the vector of coefficients,
  'aux' is the auxiliary paratmeter for the model, 'mean_PPD' is the mean prediction using
  the response scale, 'lp_' is the performance metric.
}
\description{
This internal function runs a frequentist model and returns the results.
}
