% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_info.r
\name{parse_formula}
\alias{parse_formula}
\title{Parse a formula into its component parts.}
\usage{
parse_formula(formula, data)
}
\arguments{
\item{formula}{A formula in the format 'outcome ~ variables + specials' with all special components.}

\item{data}{The raw data that the formula will be run on.}
}
\value{
A list with the transformed formula, the transformed data (based on functions in the formula),
original data, the originak formula, the response variables, the explanatory variables, the special
components to add to the formula, the random effects, and the fixed effects.
}
\description{
This internal function takes a formula and data, parses the various components of the formula,
formats the data, and returns everything as a list.
}
\seealso{
Other model setup functions: 
\code{\link{determine_model}()},
\code{\link{identify_tve}()},
\code{\link{produce_model_function}()}
}
\concept{model setup functions}
