% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{pr_list}
\alias{pr_list}
\title{Function to easily create a nicely formatted prediction list.}
\usage{
pr_list(..., .constant = NULL, .diff.in.diff = NULL, .add = NULL)
}
\arguments{
\item{...}{Variables that will be used to create the predictions.}

\item{.constant}{Variable(s) to hold constant within a contrast.}

\item{.diff.in.diff}{Variable(s) to use to create a "difference-in-difference" contrast.}

\item{.add}{Text to add to a contrast.}
}
\description{
This function allows you to create formatted predictions and contrasts.
}
\examples{
pr_list(treatment = c(1, 0), condition = c("High", "Low"), .constant = "condition")

}
\keyword{contrast}
\keyword{prediction,}
