% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_info.r
\name{produce_model_function}
\alias{produce_model_function}
\title{Produce an object with model-specific functions and arguments.}
\usage{
produce_model_function(
  model.type,
  formula.parsed,
  inference,
  model.extra.args = NULL,
  main.ivs = NULL
)
}
\arguments{
\item{model.type}{The type of model to run.}

\item{formula.parsed}{The parsed formula object.}

\item{inference}{Whether the model should be run using 'frequentist' or 'Bayesian' inference.}

\item{model.extra.args}{Extra arguments passed to the function used to run the model.}

\item{main.ivs}{A list of the main effects (e.g., a treatment) that the analysis is focused on. Used on
to identify if a survival model has time-varying hazards.}
}
\value{
Based on the type of model to run and the statistical inference used, this function returns a list of
model-specific functions and flags that allow the model to be easily run.

The functions model_run.frequentist' and 'model_run.bayesian' take the returns from this function and actually run the model.
For a frequentist approach the model is run once for each random sample of the main data. For the Bayesian approach
the model is run once for each chain.
}
\description{
This internal function takes the model type, parsed formula, inference, and extra args and returns
a list with all of the specific information (e.g., extra functions) needed to run the model. The output
is used in the frequentist and Bayesian internal model run functions.
}
\details{
For linear and categorical regression, the performance metric is log-posterior. For survival models concordance is used.
For survival models, mean_PPD is the occurrence of the outcome minus the residual.
}
\seealso{
Other model setup functions: 
\code{\link{determine_model}()},
\code{\link{identify_tve}()},
\code{\link{parse_formula}()}
}
\concept{model setup functions}
