% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{remove_backticks}
\alias{remove_backticks}
\title{Remove backticks.}
\usage{
remove_backticks(str)
}
\arguments{
\item{str}{The string to remove backticks from.}
}
\value{
A string with backticks removed.
}
\description{
This function takes in a string and outputs a string with
backticks removed.
}
