% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{resample_df}
\alias{resample_df}
\title{Function to resample a dataframe that takes into account user-specified weights and clusters.}
\usage{
resample_df(data, n = 1000, cluster = NULL, weights = NULL)
}
\arguments{
\item{data}{Dataframe with variables to be resampled}

\item{n}{Number of resamples to generate.}

\item{cluster}{Formula specifying the variable in data that identifies clusters within the data.}

\item{weights}{Vector of weights to use.}
}
\description{
This function allows you to resample a dataframe for analysis.
}
\examples{
resample_df(data, n = 2000, cluster = ~ cluster, weights = data$.weights)

}
\keyword{bootstrap,}
\keyword{cluster,}
\keyword{resample,}
\keyword{weight}
