% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results}
\alias{results}
\title{Function to get results from an object returned by analysis.}
\usage{
results(
  object,
  predictions = NULL,
  method = c("observed values", "mean"),
  times = NULL,
  draws = 1000,
  .full.matrix = F
)
}
\arguments{
\item{object}{Object returned from 'analysis' function.}

\item{predictions}{The predictions returned from "pr_list."}

\item{method}{Method for producing predictions Either uses observed values (the default) or mean values.}

\item{times}{Vector of times to produce predictions for. Only used for survival analysis.}

\item{.full.matrix}{Whether to return structured predictions (the default) or the full matrix.}
}
\value{
A list with coefficient values, predictions, and contrasts. If '.full.matrix' is set, the unsummarized matrix
  for both predictions and contrasts is returned as well.
}
\description{
This function allows you to get results based on specified predictions from the return object produced by the
'analyze' function. Results are produced on the scale of the response variable.
}
\details{
Predictions estimate the value of the response variable when one or more independent variables is set to a desired
value (as indicated by 'predictions'). Contrasts compare two or more predictions.
}
\examples{
results(object = output, predictions = main.predictions)

}
\seealso{
Other results: 
\code{\link{qualitative_assessment}()},
\code{\link{results_mediation}()},
\code{\link{summarize_interval}()}
}
\concept{results}
\keyword{bootstrap}
\keyword{prediction}
\keyword{results}
