% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results_coefficients}
\alias{results_coefficients}
\title{Produce coefficients from a frequentist or Bayesian object.}
\usage{
results_coefficients(obj.formula, obj.draws, obj.data, times = NULL)
}
\arguments{
\item{obj.formula}{The formula used in the model. A return from 'get_formula'.}

\item{obj.draws}{The draws produced from the model. A return from 'get_draws'.}

\item{obj.data}{The data used in the model. A return from 'get_data'.}

\item{times}{An optional vector identifying the times to produce coefficients for. Only relevant for
survival models.}
}
\value{
The coefficient table.
}
\description{
This function takes the draws result from a model and produces a coefficient table with confidence
intervals and P-values.
}
\seealso{
Other internal functions: 
\code{\link{get_coef}()},
\code{\link{get_data}()},
\code{\link{get_draws}()},
\code{\link{get_formula}()},
\code{\link{get_link}()},
\code{\link{get_vcov}()},
\code{\link{results_contrasts}()},
\code{\link{results_predictions}()},
\code{\link{structure_predictions}()}
}
\concept{internal functions}
