% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results_contrasts}
\alias{results_contrasts}
\title{Produce contrasts from a frequentist or Bayesian object.}
\usage{
results_contrasts(pr, predict.df)
}
\arguments{
\item{pr}{The structured predictions. A return from 'structure_predictions'.}

\item{predict.df}{The predictions produced from the model. A return from 'results_predictions'.}
}
\value{
The contrasts table.
}
\description{
This function takes the draws result from a model and produces a contrast table with confidence
intervals and P-values based on structured predictions. Contrasts are comparisons between
predictions.
}
\details{
Supports a two-way and a four-way comparison. A two-way comparison compares two different predictions
while a four-way comparison compares the difference between two two-way comparisons. For instance, identifying
a treatment effect is a two-way comparison (high vs. low value). Comparing the treatment effect in two different
situations is a four way comparison (high vs. low in situation one compared to high vs. low in situation two).
}
\seealso{
Other internal functions: 
\code{\link{get_data}()},
\code{\link{get_draws}()},
\code{\link{get_formula}()},
\code{\link{results_coefficients}()},
\code{\link{results_predictions}()},
\code{\link{structure_predictions}()}
}
\concept{internal functions}
