% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results_mediation}
\alias{results_mediation}
\title{Function to run a mediation analysis.}
\usage{
results_mediation(
  m.mediator,
  m.outcome,
  predictions,
  times = NULL,
  draws = 1000,
  .outcome = NULL
)
}
\arguments{
\item{m.mediator}{Object returned from 'analysis' function that identifies the effect of the treatment on the mediator.}

\item{m.outcome}{Object returned from 'analysis' function that identifies the effect of the treatment and the mediator on the outcome.}

\item{predictions}{A 'pr_list' object with the desired change in the treatment. The treatment should be in both equations.}

\item{.outcome}{Optional name for the outcome. If this is missing the name of the variable is used automatically.}
}
\description{
This function allows you to run mediation analysis based on two sets of returns from the 'analysis' function.
}
\examples{
results_mediation(m.mediator, m.outcome, predictions = predictions.mediation)

}
\keyword{bootstrap}
\keyword{mediation}
\keyword{results}
