% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results_predictions}
\alias{results_predictions}
\title{Produce predictions from a frequentist or Bayesian object.}
\usage{
results_predictions(
  object,
  pr,
  obj.draws,
  method = c("observed values", "mean"),
  draws = 1000,
  times = NULL,
  m = NULL
)
}
\arguments{
\item{object}{The model object.}

\item{pr}{The structured predictions. A return from 'structure_predictions'.}

\item{obj.draws}{The draws produced from the model. A return from 'get_draws'.}

\item{method}{A string indicating how variables not in 'predictions' should be treated. There are two options; "observed values" and "mean."}

\item{times}{An optional vector identifying the times to produce coefficients for. Only relevant for
survival models.}

\item{m}{The model to get results for. Only relevated for a multivariate Bayesian model (a model with
multiple dependent variables).}
}
\value{
The predictions table.
}
\description{
This function takes the draws result from a model and produces a prediction table with confidence
intervals and P-values based on structured predictions.
}
\seealso{
Other internal functions: 
\code{\link{get_data}()},
\code{\link{get_draws}()},
\code{\link{get_formula}()},
\code{\link{get_inv_link}()},
\code{\link{get_vcov}()},
\code{\link{results_coefficients}()},
\code{\link{results_contrasts}()},
\code{\link{structure_predictions}()}
}
\concept{internal functions}
