% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{run_all_replicates}
\alias{run_all_replicates}
\title{Run all replicates}
\usage{
run_all_replicates(
  replicates,
  boot_function,
  N = length(replicates),
  args = NULL,
  max.threads = 20,
  parallel.libraries = NULL
)
}
\arguments{
\item{replicates}{A list that contains a value to be passed to the 'boot_function' for each run.}

\item{boot_function}{An arbitrary function that is run once for each item in the 'replicates' list.
The value in each list item in 'replicates' is passed as the first value to 'boot_function'.}

\item{N}{The number of runs to do, which defaults to the length of 'replicates'.}

\item{args}{Additional values to pass to 'boot_function'.}

\item{max.threads}{The maximum number of threads to use. Will attempt to use the smaller of 75% of available
threads or 'max.threads'.}

\item{parallel.libraries}{A list of libraries that are loaded for each thread.}
}
\description{
This function makes it easy to run an arbitrary function in parallel.
}
\keyword{bootstrap}
\keyword{resample}
