% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{structure_predictions}
\alias{structure_predictions}
\title{Structure predictions to make testing easier.}
\usage{
structure_predictions(formula, data, predictions, method)
}
\arguments{
\item{formula}{The formula used to identify all variables that are in the model that will be run.}

\item{data}{The data used to create predictions.}

\item{predictions}{The predictions to test. This should be the object returned from 'pr_list'.}

\item{method}{A string indicating how variables not in 'predictions' should be treated. There are two options; "observed values" and "mean."}
}
\value{
A dataframe
}
\description{
This function structures data for testing a prediction. Deals with predictions that are a matrix for mediation analysis. Also
properly format factor variables.
}
\seealso{
Other internal functions: 
\code{\link{get_data}()},
\code{\link{get_draws}()},
\code{\link{get_formula}()},
\code{\link{get_inv_link}()},
\code{\link{get_vcov}()},
\code{\link{results_coefficients}()},
\code{\link{results_contrasts}()},
\code{\link{results_predictions}()}
}
\concept{internal functions}
