% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{summarize_interval}
\alias{summarize_interval}
\title{Function to summarize a dataframe and produce confidence intervals.}
\usage{
summarize_interval(
  data,
  .ci = 0.95,
  .grouping = c(".prediction.id", "prob"),
  .round = 3
)
}
\arguments{
\item{data}{Dataframe with variables to be summarized.}

\item{.ci}{Confidence interval. Defaults to 95 percent, which is the 2.5th to the 97.5th percentile of the distribution.}

\item{.grouping}{Variables that are excluded from the group and summarize. The variable "prob" should contain the estimates.}

\item{.round}{The number of digits to round the results to. Can be set to 'NULL' to prevent rounding.}
}
\description{
This function summarizes predictions to create confidence intervals.
}
\examples{
summarize_interval(results)

}
\keyword{ci}
\keyword{confidence_interval}
\keyword{prediction}
\keyword{summarize}
