% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{symlog}
\alias{symlog}
\alias{symsqrt}
\title{Symmetric 'log1p'.}
\usage{
symlog(x)

symsqrt(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The transformed numeric vector.
}
\description{
This function takes in a positive or negative number and returns correctly
transformed values.
}
