% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{trim_formula}
\alias{trim_formula}
\title{Identify variables that correlate with our response and modify the formula.}
\usage{
trim_formula(
  formula,
  data,
  cluster = NULL,
  inference = "frequentist",
  threshold = 0.4,
  save.re = F
)
}
\arguments{
\item{formula}{Formula to examine.}

\item{data}{Data to use.}

\item{cluster}{Optional data cluster.}

\item{inference}{Type of inference to use.}

\item{threshold}{P-value threshold for inclusion or exclusion of control variables. Variables with p-values above
this threshold will be dropped.}

\item{save.re}{Whether to save the random effects in the returned formula.}
}
\value{
Returns a list with two values. The first is a new formula that only includes the variables that correlate
  with the dependent variable. The second is a list of the variables that were dropped from the formula.
}
\description{
This function allows easy identification of the control variables that correlate with the IV. This function
can be used in conjunction with 'formula_iv_as_dv' to simplify the process of taking a formula, identifying
variables in the formula that correlate with the main independent variable(s) and trimming the formula to
only include these variables.
}
\seealso{
Other research plan functions: 
\code{\link{add_transform}()},
\code{\link{analyze_plan}()},
\code{\link{build_other_values}()},
\code{\link{create_new_formula}()},
\code{\link{create_plan_data}()},
\code{\link{create_research_formulas}()},
\code{\link{create_variable_table}()},
\code{\link{formula_iv_as_dv}()},
\code{\link{identify_problematic_variables}()},
\code{\link{identify_variable_transformations}()},
\code{\link{research_plan}()}
}
\concept{research plan functions}
\keyword{formula,}
\keyword{selection}
\keyword{variable,}
