% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.r
\name{analyze_plan}
\alias{analyze_plan}
\title{Function to automatically analyze a research plan in a systematic and efficient manner.}
\usage{
analyze_plan(
  research.plan,
  outcomes,
  only.run = NULL,
  run.basic = T,
  run.mediation = T,
  run.interaction = T
)
}
\description{
Function to automatically analyze a research plan in a systematic and efficient manner.
}
