% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{formula_iv_as_dv}
\alias{formula_iv_as_dv}
\title{Reformulates the formula so that the IV is the DV.}
\usage{
formula_iv_as_dv(x)
}
\arguments{
\item{x}{Object of class 'planformula'.}
}
\description{
This function is a helper function to allow easy identification of the controls that correlate with the IV.
}
\keyword{formulas,}
\keyword{variables}
