% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{parse_contrast}
\alias{parse_contrast}
\title{Helper function to parse the 'contrasts' string in a result.}
\usage{
parse_contrast(df)
}
\description{
Helper function to parse the 'contrasts' string in a result.
}
