% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{print.planformula}
\alias{print.planformula}
\alias{show.planformula}
\title{Print and show methods for planformula}
\usage{
\method{print}{planformula}(x, useS4 = F)

\method{show}{planformula}(x)
}
\description{
Print and show methods for planformula
}
