% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{qualitative_assessment}
\alias{qualitative_assessment}
\title{Provide a qualitative assessment for a set of results (main, interaction, and mediation effects)}
\usage{
qualitative_assessment(research.plan, all.results)
}
\description{
Provide a qualitative assessment for a set of results (main, interaction, and mediation effects)
}
