% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{research_plan}
\alias{research_plan}
\title{Function to get create new variables and formulas for testing a set of variable relationships.}
\usage{
research_plan(
  treatment,
  control = NULL,
  interaction = NULL,
  mediation = NULL,
  data,
  keep.vars = NULL,
  factor.transform = NULL,
  prefix = list(source = "sv", target = "tv", other = "oth"),
  unit = list(unit = NULL, source = NULL, target = NULL, time = NULL),
  lag = NULL
)
}
\arguments{
\item{treatment}{Variable containing names of treatments to test. Can be a vector for just actions or a list that contains both 'action' and 'factor' to test.}

\item{control}{Optional vector containing variable names for controls to include.}

\item{interaction}{Optional vector containing interactions for the treatment.}

\item{mediation}{Optional vector containing mediating variables for the treatment.}

\item{data}{Data containing variables in treatment, control, interaction, and mediation.}

\item{keep.vars}{Optional vector of variable names to keep in the data frame returned.}

\item{prefix}{List of prefixes for source and target in data as well as for "other" treatments used against source/target. Source and target are optional.}

\item{unit}{List of variables that correspond to units, times, and optional source/target in data.}

\item{lag}{Optional vector containing the number of lags to use.}
}
\description{
This function produces a set of formulas and new data for testing a research plan.
}
\examples{
research_plan(treatment = c("use.force", "show.force"),
control = c("wdi.economy", "wdi.terrain"), data = data,
unit = list(unit = "dispute.id", time = "time"), lag = 1:2)

}
\keyword{formulas,}
\keyword{variables}
