% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.r
\name{results}
\alias{results}
\title{Function to get results from an object returned by analysis.}
\usage{
results(
  object,
  predictions = NULL,
  method = c("observed values", "mean"),
  times = NULL,
  draws = 1000,
  .full.matrix = F
)
}
\arguments{
\item{object}{Object returned from 'analysis' function.}

\item{predictions}{The predictions returned from "pr_list."}

\item{method}{Method for producing predictions Either uses observed values (the default) or mean values.}

\item{times}{Vector of times to produce predictions for. Only used for survival analysis.}

\item{.full.matrix}{Whether to return structured predictions (the default) or the full matrix.}
}
\description{
This function allows you to get results from an analysis.
}
\examples{
results(object = output, predictions = main.predictions)

}
\keyword{bootstrap}
\keyword{prediction}
\keyword{results}
