% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.r
\name{trim_formula}
\alias{trim_formula}
\title{Identify variables that correlate with our response and modify the formula.}
\usage{
trim_formula(
  formula,
  data,
  cluster = NULL,
  inference = "frequentist",
  threshold = 0.4,
  save.re = F
)
}
\arguments{
\item{formula}{Formula to examine.}

\item{data}{Data to use.}

\item{cluster}{Optional data cluster.}

\item{inference}{Type of inference to use.}

\item{threshold}{P-value threshold.}

\item{save.re}{Whether to save the random effects in a formula.}
}
\description{
This function allows easy identification of the controls that correlate with the IV.
}
\keyword{formula,}
\keyword{selection}
\keyword{variable,}
